feat: swagger 보안 추가
