+++ backend/.format.env
@@ -16,3 +16,5 @@ S3_ENDPOINT=
 S3_BUCKET=
 CONTENT_PREFIX=
 GOOGLE_APPLICATION_CREDENTIALS=
+SWAGGER_ID=
+SWAGGER_PASSWORD=
